fix: made upload and download work through CLI
