+++ src/api/discord/download.js
@@ -1,7 +1,10 @@
 require('dotenv').config();
 const {Client, IntentsBitField} = require('discord.js');
+const mongoose = require('mongoose');
+const Thread = require('../../model/thread');
 
-function download(text) {
+// the format to download is node src/index.js -d "title"
+async function download(title) {
     
     // Create a new client instance with intents (basically set of permissions that the bot needs)
     const client = new Client({
@@ -12,17 +15,41 @@ function download(text) {
             IntentsBitField.Flags.MessageContent
         ]
     });
+
+    mongoose.connect('mongodb://localhost:27017/discord').then(() => {
+        console.log("connected to mongo")
+    }). catch(err => {
+        console.log(err)
+    })
     
     client.on('ready', async () => {
         console.log(`Logged in as ${client.user.tag}!`);
-        const guild = client.guilds.cache.get(process.env.GUILD_ID);
-        const channel = guild.channels.cache.get(process.env.CHANNEL_ID);
-        const thread = channel.threads.cache.get(process.env.THREAD_ID);
-        const messages = await thread.messages.fetch();
-        messages.forEach(msg => {
-            console.log(`${msg.author.tag}: ${msg.content}`);
-        });
+
+        try {
+            // fetching a thread from the database
+            const res = await Thread.findOne({
+                NAME: title
+            })
+    
+            const guild = client.guilds.cache.get(res.GUILD_ID);
+            const channel = guild.channels.cache.get(res.CHANNEL_ID);
+            const thread = channel.threads.cache.get(res.THREAD_ID);
+    
+            // fetching the messages from the thread using guild id, channel id and thread id
+            const messages = await thread.messages.fetch();
+    
+            // reversing as the latest message will be featched first
+            const reversedMessages = messages.reverse();
+            reversedMessages.forEach(msg => {
+                console.log(`${msg.author.tag}: ${msg.content}`);
+            });
+        } catch (error) {
+            console.log(error)
+        }
+        await client.destroy();
     });
+
+    client.login(process.env.LOGIN_TOKEN);
 }
 
-export default download;
+module.exports = download
